# api/openapi.yaml
openapi: 3.0.3
info:
  title: NGX-CP API
  description: |
    Nginx 集群管理平台（NGX-CP）控制面 API。

    当前处于 **M1 骨架**阶段，已实现：
    - 节点域 CRUD（node）与一次性接入令牌（enroll token）
    - Bearer Token 鉴权（写接口需带 `Authorization: Bearer <token>`，未配置 token 时所有写接口返回 401）
    - 审计中间件（仅记录成功的写操作到 audit_logs，best-effort）
    - 能力基线占位（真实 `nginx -V` / `-T` 解析在 T016/T017 落地）

    后续 M2+ 会逐步补齐配置中心、发布流水线、证书、LVS、日志、预警等。

    本文档的作用：
    - 在 **Apifox** 中「项目 → 导入 → OpenAPI / Swagger」选择本文件，即可生成可调试的 API 文档与 Mock；
    - 作为前后端（Vue3 前端 / Go 控制面）的接口契约，随代码版本一起演进。
      新增/变更接口时更新本文件，并在 Apifox 用「覆盖导入」刷新。
  version: 0.2.0
  contact:
    name: th815
    url: https://github.com/th815/NGX-CP
servers:
  - url: http://localhost:8080
    description: 本地开发（go run ./cmd/ngxcp-server，默认监听 :8080）
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/nodes:
    get:
      tags: [Cluster]
      summary: 节点列表（支持 role / status 过滤与分页）
      operationId: listNodes
      security: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [real_server, director, director_and_rs, unknown]
        - name: status
          in: query
          schema:
            type: string
            enum: [online, offline, degraded, enrolling, decommissioned]
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: 节点列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeList'
    post:
      tags: [Cluster]
      summary: 新建节点（写接口，需鉴权）
      operationId: createNode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeCreate'
      responses:
        '200':
          description: 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeNode'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/nodes/{id}/join-token:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    post:
      tags: [Cluster]
      summary: 为指定节点轮换（重新签发）Join Token（写接口，需鉴权）
      description: >
        令牌过期或需吊销旧令牌时调用。先吊销该节点所有未吊销的 Join Token（即时生效），
        再签发新令牌；与 /join-token/revoke（只吊销不签发）互补。
      operationId: rotateJoinToken
      parameters:
        - name: ttl
          in: query
          description: 新令牌有效期，Go duration 格式（如 24h），默认 24h
          schema: { type: string }
      responses:
        '200':
          description: 已轮换；返回新令牌（仅此返回一次）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/nodes/{id}/join-token/revoke:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    post:
      tags: [Cluster]
      summary: 吊销该节点所有尚未吊销的 Join Token（写接口，需鉴权）
      description: >
        与 rotateJoinToken 互补，仅吊销不签发。用于令牌疑似泄漏或节点下线等安全事件响应；
        吊销即时生效，Agent 持已被吊销令牌注册将被拒绝（CodeUnauthorized）。
      operationId: revokeJoinToken
      responses:
        '200':
          description: 已吊销该节点的全部活跃 Join Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
              example:
                code: 0
                message: ok
                data:
                  revoked: 1
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Bearer Token（由配置 `auth_admin_token` / 环境变量 `NGXCP_AUTH_ADMIN_TOKEN` 设定）。
        留空时所有写接口返回 401（fail-closed）。本地开发可用 `NGXCP_AUTH_ADMIN_TOKEN=dev-admin-token go run ./cmd/ngxcp-server`。
  responses:
    BadRequest:
      description: 入参非法（如未知 role / 缺字段）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
          example:
            code: 4001
            message: 非法 role
    Unauthorized:
      description: 缺少 / 错误的鉴权令牌
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
          example:
            code: 4001
            message: 未授权
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
          example:
            code: 4004
            message: 节点不存在
  schemas:
    Envelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 业务码，0 表示成功（非 0 见 docs/ARCHITECTURE 错误码表）
          example: 0
        message:
          type: string
          example: ok
        detail:
          type: string
          description: 失败时的明细（成功时省略），超长会被截断
        data:
          type: object
          nullable: true
          description: 业务负载；结构随接口而异
    Node:
      type: object
      required: [id, name, address, role, status, lvs_weight, lvs_enabled]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: nx-rs-01
        address:
          type: string
          example: 10.0.0.11:9443
        role:
          type: string
          enum: [real_server, director, director_and_rs, unknown]
          example: real_server
        status:
          type: string
          enum: [online, offline, degraded, enrolling, decommissioned]
          example: enrolling
        lvs_weight:
          type: integer
          example: 100
        lvs_enabled:
          type: boolean
          example: false
        last_heartbeat_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NodeCreate:
      type: object
      required: [name, address]
      properties:
        name:
          type: string
          example: nx-rs-01
        address:
          type: string
          example: 10.0.0.11:9443
        role:
          type: string
          enum: [real_server, director, director_and_rs, unknown]
          description: 留空则入库为 unknown，待能力上报后由 T019 识别
        lvs_weight:
          type: integer
          example: 100
...
    EnvelopeNode:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Node'
    NodeList:
      type: object
      required: [code, message, data, total]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          type: array
          items:
            $ref: '#/components/schemas/Node'
        total:
          type: integer
          example: 1